[IT]Optimize the class naming.
